fix all compiler errors

+++ after/egrep.c
@@ -1619,7 +1619,7 @@ yylex(void) {
 			}
 			do {
 				if (c == '\0' || c == EOF) synerror();
-				if (nxtchar >= MAXCHARS) more_chars(nxtchar);
+				if ((unsigned)nxtchar >= MAXCHARS) more_chars(nxtchar);
 				chars[nxtchar++] = c;
 				cclcnt++;
 			} while ((c = nextch()) != ']');
@@ -1627,10 +1627,11 @@ yylex(void) {
 			return (x);
 		case '\\':
 			if ((c = nextch()) == EOF) synerror();
+			/* fall through */
 		defchar:
 		default:
 			if (mbcode && c & ~(wchar_t)0177) {
-				if (nxtchar >= MAXCHARS) more_chars(nxtchar);
+				if ((unsigned)nxtchar >= MAXCHARS) more_chars(nxtchar);
 				chars[nxtchar] = c;
 				yylval = nxtchar++;
 				return (MCHAR);
@@ -1646,7 +1647,7 @@ synerror(void) {
 
 static int
 enter(int x) {
-	if(line >= MAXLIN) more_lines(line);
+	if((unsigned)line >= MAXLIN) more_lines(line);
 	name[line] = x;
 	left[line] = 0;
 	right[line] = 0;
@@ -1671,7 +1672,7 @@ menter(int x) {
 
 static int
 node(int x, int l, int r) {
-	if(line >= MAXLIN) more_lines(line);
+	if((unsigned)line >= MAXLIN) more_lines(line);
 	name[line] = x;
 	left[line] = l;
 	right[line] = r;
@@ -1682,7 +1683,7 @@ node(int x, int l, int r) {
 
 static int
 unary(int x, int d) {
-	if(line >= MAXLIN) more_lines(line);
+	if((unsigned)line >= MAXLIN) more_lines(line);
 	name[line] = x;
 	left[line] = d;
 	right[line] = 0;
@@ -1715,7 +1716,7 @@ cgotofn(int s, int c)
 	int curpos, num;
 	int number, newpos;
 	n = lastn;
-	cc = iflag ? mbcode && c & ~(wchar_t)0177 ? towlower(c):tolower(c) : c;
+	cc = iflag ? mbcode && c & ~(wchar_t)0177 ? (int)towlower(c):tolower(c) : c;
 	num = positions[state[s]];
 	count = icount;
 	for (i=3; i <= line; i++) tmpstat[i] = initstat[i];
@@ -1726,11 +1727,11 @@ cgotofn(int s, int c)
 			if (
 				((cc & ~(wchar_t)(NCHARS-1)) == 0 && k == cc)
 				|| (k == MCHAR && cc == chars[right[curpos]])
-				|| (cc != '\n' && cc != WEOF && ((k == DOT)
+				|| (cc != '\n' && cc != (int)WEOF && ((k == DOT)
 				|| (k == CCL && member(cc, right[curpos], 1))
 				|| (k == NCCL && member(cc, right[curpos], 0))))
 			) {
-				if (foll[curpos] < MAXPOS)
+				if ((unsigned)foll[curpos] < MAXPOS)
 					number = positions[foll[curpos]];
 				else
 					number = 0;
@@ -1824,7 +1825,7 @@ member(int symb, int set, int torf) {
 	pos = set + 1;
 	num += pos;
 	while (pos < num) {
-		if (chars[pos] == '-' && lastc != WEOF) {
+		if (chars[pos] == '-' && (unsigned)lastc != WEOF) {
 			if (++pos == num)	/* System V oddity: '[a-]' */
 				return (!torf);	/* matches 'a' but not '-' */
 			if (symb > lastc && symb < chars[pos])
@@ -1861,7 +1862,7 @@ add(int *array, int n) {
 	positions[nxtpos++] = count;
 	for (i=3; i <= line; i++) {
 		if (tmpstat[i] == 1) {
-			if (nxtpos >= MAXPOS) more_positions(nxtpos);
+			if ((unsigned)nxtpos >= MAXPOS) more_positions(nxtpos);
 			positions[nxtpos++] = i;
 		}
 	}
